"form manager"

+++ src/main/java/com/gxu/informationLibrary/serviceImpl/tbManageImpl.java
@@ -40,7 +40,10 @@ public class tbManageImpl implements tbManageServer {
     }
 
     @Override
-    public void deleteTable(String db_name, String tb_name) {
+    public void deleteTable(String db_name, String tb_name) throws Exception {
+        if (tbManage.count(db_name,tb_name)>0){
+            throw new Exception("数据库表非空，不允许删除");
+        }
         tbManage.deleteTable(db_name, tb_name);
     }
